extrect struct and fixed values to config
